Fixing everything except editing
